RP USBv1: avoid 16-bit truncation of transfer counts
usb_serve_endpoint() computed the remaining IN transfer size
txsize - txcnt into a uint16_t: a transfer larger than 64 KiB was
truncated to a wrong remainder and a transfer whose remainder was an
exact multiple of 65536 terminated early with data still to send.
txsize and txcnt are size_t, make n size_t as well (the OUT use of n is
a hardware transfer length of at most 1023 bytes and is unaffected).

Widen rxpkts to size_t too: it is derived from the size_t rxsize and
the uint16_t counter capped OUT transfers at 65535 packets, and drop
the now pointless narrowing cast in usb_lld_start_out().

The buf_len locals in the buffer preparation functions stay uint16_t,
they are bounded by the clamp to the endpoint's max packet size (at
most 1023); comments now say so.

+++ os/hal/ports/RP/LLD/USBv1/hal_usb_lld.h
@@ -180,8 +180,10 @@ typedef struct {
   /* End of the mandatory fields.*/
   /**
    * @brief   Number of packets to receive.
+   * @note    Derived from the @p size_t rxsize, a 16-bit counter would
+   *          truncate large transfers.
    */
-  uint16_t                      rxpkts;
+  size_t                        rxpkts;
   /**
    * @brief   Data PID used by next transfer.
    */
